+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ```
+* Commit: 605a2a5 - v3
+| Author: FJrodafo - Sat, 22 Jun 2024 14:02:17 +0100
+| 
 * Commit: 3a25d2b - v2
 | Author: FJrodafo - Sat, 22 Jun 2024 12:31:34 +0100
 |